+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - add: modality support
 - add: folder based on db table prefix
 - add: support for parquet files and Delta/Hive/Iceberg directories
-- add: support for SAS files
+- add: support for SAS, SPSS and Stata files
 
 ## 0.2.1 (2026-01-21)